Code improvements

+++ src/Common.cpp
@@ -54,7 +54,7 @@ namespace NWScriptPluginCommons {
 
     // Create a thousand separator string
     // Extracted from: https://stackoverflow.com/questions/43482488/how-to-format-a-number-with-thousands-separator-in-c-c
-    std::string thousandSeparator(int number)
+    std::string thousandSeparator(long long number)
     {
         auto count = 3;
         std::string src = std::to_string(number);
@@ -81,7 +81,7 @@ namespace NWScriptPluginCommons {
 
     // Create a thousand separator string
     // Extracted from: https://stackoverflow.com/questions/43482488/how-to-format-a-number-with-thousands-separator-in-c-c
-    generic_string thousandSeparatorW(int number)
+    generic_string thousandSeparatorW(long long number)
     {
         return str2wstr(thousandSeparator(number));
     }
